Tweak icon fills
Use same fills for incognito and other icons.
Improve look of dark shortcut picker.
Add dark version of clear icon so we can use it as a background image.  There doesn't seem to be a way to use CSS to affect the fills of background svg images.
Get rid of inline width/height on the svg icons and use styles.
Fix default color of help icon in light mode.